Update Makefile and README.md
This commit updates the Makefile and README.md files to include the latest dependencies for prettier, eslint, and langchain-visualizer. The main.py file has been updated to allow for an external objective to be passed as input. The AgentOrchestrator class has been added to orchestrate different modules to achieve the given objective.